sem2/DSA5203/Assignment_1/requirements.txt
numpy
Pillow
matplotlib
scipy